re-added group availability pull from group
