+++ src/components/GroupAvailability.jsx
@@ -10,6 +10,7 @@ import Draggable from 'react-draggable';
 
 import { collection, getDocs } from "firebase/firestore";
 import { db } from '../firebase';
+import { fetchGroupAvailability } from '../utils/fetchGroupData';
 
 const buttonTheme = createTheme({
   palette: {
@@ -42,47 +43,16 @@ function makeSlotKey(date, hourLabel, isHalfHour = false) {
   return `${iso} ${hourLabel}:${minutes}`;
 }
 
-function GroupSchedule({ startTime, endTime, startDate, endDate }) {
+function GroupSchedule({ groupData, startTime, endTime, startDate, endDate }) {
   const [selectedBlocks, setSelectedBlocks] = useState(new Set());
   const [isMouseDown, setIsMouseDown] = useState(false);
   const [showPopup, setShowPopup] = useState(false);
   const [groupAvailabilityData, setGroupAvailabilityData] = useState({});
-  const [numMembers, setNumMembers] = useState(0);
   const [lockedDate, setLockedDate] = useState(null);
+  const numMembers = groupData.participants?.length || 0;
 
   useEffect(() => {
-    async function fetchAvailabilityData() {
-      const data = {};
-      const querySnapshot = await getDocs(collection(db, "availability"));
-      let members = 0;
-
-      querySnapshot.forEach((doc) => {
-        members++;
-        const userID = doc.id;
-        for (const date in doc.data()) {
-          const slots = doc.data()[date]['data'];
-          if (slots === undefined) continue;
-          // console.log(date, slots);
-
-          const compressedSlots = [];
-          for (let i = 0; i < slots.length; i += 2) {
-            const group = slots.slice(i, i + 2);
-            compressedSlots.push(group.every(slot => slot === 1) ? 1 : 0);
-          }
-
-          if (date in data) {
-            data[date] = data[date].map((num, index) => num + compressedSlots[index]);
-          } else {
-            data[date] = compressedSlots;
-          }
-        }
-      });
-
-      setGroupAvailabilityData(data);
-      setNumMembers(members);
-    }
-
-    fetchAvailabilityData();
+    setGroupAvailabilityData(fetchGroupAvailability(groupData));
   }, []);
 
   // Generate all dates in range
@@ -431,7 +401,7 @@ function formatYyyyMmDd(date) {
   return `${y}-${m}-${d}`;
 }
 
-export default function GroupAvailability({ startDate, endDate, startTime, endTime}) {
+export default function GroupAvailability({ groupData, startDate, endDate, startTime, endTime}) {
   const [weekStart, setWeekStart] = useState(() => new Date(startDate));
   const [weekEnd, setWeekEnd] = useState(() => new Date(endDate));
 
@@ -481,6 +451,7 @@ export default function GroupAvailability({ startDate, endDate, startTime, endTi
       </div>
 
       <GroupSchedule
+        groupData={groupData}
         startTime={startTime}
         endTime={endTime}
         startDate={formatYyyyMmDd(weekStart)}
